+++ cedar-policy/CHANGELOG.md
@@ -35,6 +35,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improve parser error messages to more reliably notice that a function or
+  method does exists when it is called with an incorrect number of arguments or
+  using the wrong call style.
 - Include source spans on more parser error messages.
 - Better integration with `miette` for various error types. If you have
   previously been just using the `Display` trait to get the error message from a